+++ opennebula/resource_opennebula_virtual_machine_nic_test.go
@@ -52,7 +52,21 @@ func TestAccVirtualMachineNICUpdate(t *testing.T) {
 					resource.TestCheckResourceAttr("opennebula_virtual_machine.test", "name", "test-virtual_machine"),
 					resource.TestCheckResourceAttr("opennebula_virtual_machine.test", "nic.#", "2"),
 					resource.TestCheckResourceAttr("opennebula_virtual_machine.test", "nic.0.computed_ip", "172.16.100.131"),
+					resource.TestCheckResourceAttr("opennebula_virtual_machine.test", "nic.0.computed_gateway", "172.16.100.1"),
+					resource.TestCheckResourceAttr("opennebula_virtual_machine.test", "nic.0.computed_dns", "8.8.8.8"),
 					resource.TestCheckResourceAttr("opennebula_virtual_machine.test", "nic.1.computed_ip", "172.16.100.112"),
+				),
+			},
+			{
+				Config: testAccVirtualMachineTemplateConfigNICExtraUpdate,
+				Check: resource.ComposeTestCheckFunc(
+					resource.TestCheckResourceAttr("opennebula_virtual_machine.test", "name", "test-virtual_machine"),
+					resource.TestCheckResourceAttr("opennebula_virtual_machine.test", "nic.#", "2"),
+					resource.TestCheckResourceAttr("opennebula_virtual_machine.test", "nic.0.computed_ip", "172.16.100.131"),
+					resource.TestCheckResourceAttr("opennebula_virtual_machine.test", "nic.0.computed_gateway", "172.16.100.254"),
+					resource.TestCheckResourceAttr("opennebula_virtual_machine.test", "nic.0.computed_dns", "1.1.1.1"),
+					resource.TestCheckResourceAttr("opennebula_virtual_machine.test", "nic.1.computed_ip", "172.16.100.112"),
+					resource.TestCheckResourceAttr("opennebula_virtual_machine.test", "nic.1.computed_method", "skip"),
 				),
 			},
 			{
@@ -210,6 +224,8 @@ resource "opennebula_virtual_network" "network1" {
 	group = "oneadmin"
 	security_groups = [0]
 	cluster_ids = [0]
+	gateway = "172.16.100.1"
+	dns = "8.8.8.8"
   }
 
   resource "opennebula_virtual_network" "network2" {
@@ -321,23 +337,23 @@ var testAccVirtualMachineTemplateConfigNICUpdate = testNICVNetResources + `
 	  permissions = "642"
 	  memory = 128
 	  cpu = 0.1
-	
+
 	  context = {
 		NETWORK  = "YES"
 		SET_HOSTNAME = "$NAME"
 	  }
-	
+
 	  graphics {
 		type   = "VNC"
 		listen = "0.0.0.0"
 		keymap = "en-us"
 	  }
-	
+
 	  os {
 		arch = "x86_64"
 		boot = ""
 	  }
-	
+
 	  tags = {
 		env = "prod"
 		customer = "test"
@@ -352,7 +368,7 @@ var testAccVirtualMachineTemplateConfigNICUpdate = testNICVNetResources + `
 		network_id = opennebula_virtual_network.network2.id
 		ip = "172.16.100.111"
 	  }
-	
+
 	  timeout = 5
 }
 `
@@ -365,23 +381,23 @@ var testAccVirtualMachineTemplateConfigNICIPUpdate = testNICVNetResources + `
 	  permissions = "642"
 	  memory = 128
 	  cpu = 0.1
-	
+
 	  context = {
 		NETWORK  = "YES"
 		SET_HOSTNAME = "$NAME"
 	  }
-	
+
 	  graphics {
 		type   = "VNC"
 		listen = "0.0.0.0"
 		keymap = "en-us"
 	  }
-	
+
 	  os {
 		arch = "x86_64"
 		boot = ""
 	  }
-	
+
 	  tags = {
 		env = "prod"
 		customer = "test"
@@ -397,6 +413,52 @@ var testAccVirtualMachineTemplateConfigNICIPUpdate = testNICVNetResources + `
 		ip = "172.16.100.112"
 	  }
 
+	  timeout = 5
+}
+`
+
+var testAccVirtualMachineTemplateConfigNICExtraUpdate = testNICVNetResources + `
+
+  resource "opennebula_virtual_machine" "test" {
+	  name        = "test-virtual_machine"
+	  group       = "oneadmin"
+	  permissions = "642"
+	  memory = 128
+	  cpu = 0.1
+
+	  context = {
+		NETWORK  = "YES"
+		SET_HOSTNAME = "$NAME"
+	  }
+
+	  graphics {
+		type   = "VNC"
+		listen = "0.0.0.0"
+		keymap = "en-us"
+	  }
+
+	  os {
+		arch = "x86_64"
+		boot = ""
+	  }
+
+	  tags = {
+		env = "prod"
+		customer = "test"
+	  }
+
+	  nic {
+		network_id = opennebula_virtual_network.network1.id
+		ip = "172.16.100.131"
+		security_groups = [opennebula_security_group.mysecgroup.id]
+		gateway = "172.16.100.254"
+		dns = "1.1.1.1"
+	  }
+	  nic {
+		network_id = opennebula_virtual_network.network2.id
+		ip = "172.16.100.112"
+		method = "skip" # IP is assigned normally in oned, but then one-context ignores it.
+	  }
 
 	  timeout = 5
 }