+++ CHANGELOG.md
@@ -5,6 +5,7 @@ FEATURES:
 * resources/opennebula_virtual_network: allow to modify the user owning the resource (#529)
 * resources/opennebula_virtual_machine: add nil checks before type casting (#530)
 * resources/opennebula_virtual_router_nic: add floating_only nic argument (#547)
+* resources/opennebula_virtual_machine: add method, gateway, dns arguments for nics (#548)
 
 ENHANCEMENTS: